feat: extract CreateAuditingEntity from AuditingEntity
