+++ backend/src/main/java/wooteco/prolog/common/entity/CreateAuditingEntity.java
@@ -1,23 +1,22 @@
-package wooteco.prolog.common;
+package wooteco.prolog.common.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
-@Getter
+import java.time.LocalDateTime;
+
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class AuditingEntity {
+public abstract class CreateAuditingEntity {
 
     @Column(nullable = false, updatable = false)
     @CreatedDate
     private LocalDateTime createdAt;
 
-    @LastModifiedDate
-    private LocalDateTime updatedAt;
+    public LocalDateTime getCreatedAt() {
+        return createdAt;
+    }
 }
