+++ backend/src/main/java/wooteco/prolog/studylog/domain/Studylog.java
@@ -13,7 +13,7 @@ import java.util.stream.Collectors;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import wooteco.prolog.common.AuditingEntity;
+import wooteco.prolog.common.entity.AuditingEntity;
 import wooteco.prolog.common.exception.BadRequestCode;
 import wooteco.prolog.common.exception.BadRequestException;
 import wooteco.prolog.member.domain.Member;
